Refactor: moved pack/unpack and tests methods to recovery.rs
