+++ chain-crypto/src/digest.rs
@@ -287,7 +287,6 @@ impl<H: DigestAlg> Digest<H> {
     }
 }
 
-use chain_ser::packer::Codec;
 use std::marker::PhantomData;
 
 /// A typed version of Digest
@@ -295,28 +294,6 @@ pub struct DigestOf<H: DigestAlg, T> {
     inner: Digest<H>,
     marker: PhantomData<T>,
 }
-
-
-fn pack_digestof<H: DigestAlg, T, W: std::io::Write>(digestof: &DigestOf<H, T>, codec: &mut Codec<W>) -> Result<(), std::io::Error> {
-    let inner_data = digestof.as_ref();
-    codec.put_u64(inner_data.len() as u64)?;
-    codec.put_bytes(inner_data)?;
-    Ok(())
-}
-
-
-fn unpack_digestof<H: DigestAlg, T, R: std::io::BufRead>(codec: &mut Codec<R>) -> Result<DigestOf<H, T>, std::io::Error> {
-    let size = codec.get_u64()?;
-    let bytes = codec.get_bytes(size as usize)?;
-    match DigestOf::try_from(&bytes[..]) {
-        Ok(data) => Ok(data),
-        Err(e) => Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            format!("{}", e),
-        )),
-    }
-}
-
 
 unsafe impl<H: DigestAlg, T> Send for DigestOf<H, T> {}
 
@@ -490,32 +467,3 @@ macro_rules! typed_define_from_instances {
 
 typed_define_from_instances!(Sha3_256, 32, "sha3");
 typed_define_from_instances!(Blake2b256, 32, "blake2b");
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use typed_bytes::ByteArray;
-
-    #[test]
-    fn test_digestof_pack_unpack_bijection() -> Result<(), std::io::Error> {
-        use std::io::Cursor;
-
-        let data: [u8; 32] = [0u8; 32];
-        let slice = &data[..];
-        let byte_array: ByteArray<u8> = ByteArray::from(Vec::from(slice));
-        let byte_slice: ByteSlice<u8> = byte_array.as_byteslice();
-        let digest: DigestOf<Blake2b256, u8> = DigestOf::digest_byteslice(&byte_slice);
-
-        let mut c: Cursor<Vec<u8>> = Cursor::new(Vec::new());
-        let mut codec = Codec::new(c);
-        pack_digestof(&digest, &mut codec)?;
-        c = codec.into_inner();
-        c.set_position(0);
-        codec = Codec::new(c);
-
-        let deserialize_digest : DigestOf<Blake2b256, u8> = unpack_digestof(&mut codec)?;
-        assert_eq!(digest, deserialize_digest);
-
-        Ok(())
-    }
-}
